third_party/blink/web_tests/platform/generic/fast/forms/reportValidity-cancel-expected.txt

Tests for reportValidity() with invalid event canceling

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

"invalid" event is not canceled.
PASS input.addEventListener("invalid", nothingListener, false); !input.reportValidity() && invalidFired is true
PASS document.activeElement is $("input-unset")
PASS invalidFired = false; !form.reportValidity() && invalidFired is true
PASS document.activeElement is $("input-unset")


"invalid" event is canceled.
PASS input.addEventListener("invalid", cancelListener, false); !input.reportValidity() && invalidFired is true
PASS document.activeElement is not $("input-unset")
PASS invalidFired = false; !form.reportValidity() && invalidFired is true
PASS document.activeElement is not $("input-unset")
PASS successfullyParsed is true

TEST COMPLETE
